delete game from detail

+++ finerleague-web/src/app/games/game-list/game-list.component.ts
@@ -23,34 +23,37 @@ export class GameListComponent implements OnInit {
     private route: ActivatedRoute,
     private router: Router,
     private spinnerService: SpinnerService
-  ) {}
+  ) { }
 
   ngOnInit() {
-      this.loggerService.log('... initializing game list component.');
-      this.spinnerService.show();
-      this.isRequesting = true;
+    this.loggerService.log('... initializing game list component.');
+    this.spinnerService.show();
+    this.isRequesting = true;
 
-      this.service.findAll()
-          .subscribe(games => {
-            this.games = games;
+    this.service.findAll()
+      .subscribe(games => {
+        this.games = games;
 
-            //TODO remove this part in real app leaving the hide of the spinner.
-            var that = this;
-            setTimeout(function(){
-              that.spinnerService.hide();
-            }, 3000);
-          }, err => {
-              // Log errors if any
-              console.log(err);
-          });
+        //TODO remove this part in real app leaving the hide of the spinner.
+        var that = this;
+        setTimeout(function() {
+          that.spinnerService.hide();
+        }, 3000);
+      }, err => {
+        // Log errors if any
+        console.log(err);
+      });
   }
 
   delete(id: string) {
-    this.service.delete(id).subscribe(res => {
-      if(res.ok) {
-        let index = this.games.findIndex(game => game.id === id);
-        this.games.splice(index, 1);
-      }
-    });
+    let confirmation = window.confirm('Are you sure you want to delete this game?');
+    if (confirmation) {
+      this.service.delete(id).subscribe(res => {
+        if (res.ok) {
+          let index = this.games.findIndex(game => game.id === id);
+          this.games.splice(index, 1);
+        }
+      });
+    }
   }
 }
